Question Tool added
When starting up the program, the user can either play the trivia maze
or manage the questions. Each type of question is managed independently.
They can be added onto, have questions removed, or viewed. The user can
then go back to the beginning and start a new maze to see the newly
added questions (maybe).

There may be an error if the user enters a string where an integer is
required. The spacing also needs some work, and the code is
unnecessarily redundant. There's areas where there's literally a one
word change, so that may be improved upon in the future.