UNITY 4.6 MUST BE INSTALLED
UNITY 4.6 MUST BE INSTALLED on your machine to be able to load this. It
isn't fully functional. The maze sets itself up depending on its size.
Basically it is how far we made in on last monday. I have run into a
dead end on getting the questions to display because of serializable. so
I am taking a break tonight so we can discuss it tomorrow.