Fixing user input
I switched the int Parsing in the Driver to a string. So, the player can
now enter "1", "u", or "up" to go up for example, as well as the other
directions and saving. Another line was added so that the player can
quit while mid-game.

Added try/catch to the int Parsing in the QuestionManager.

Checked for possible SQL Injections but couldn't find any.