Minor fixes
Synched the unit tests with the methods. The little "invalid input" bug
that occurred in class was also removed.

Spent the better portion of the night trying to get the installer to
work, but to no avail. I'd advise not bothering with the stuff that's
already there, as most of it needs to be recompiled every time anyway.